refactored wav reading/writing to use WavFile class
