+++ TestPatches/PatchRun.cpp
@@ -17,7 +17,6 @@ void vPortFree( void *pv ){
 #include "MemoryBuffer.hpp"
 #include "registerpatch.h"
 #include "WavFile.h"
-#include "wav.h"
 
 #define SAMPLE_RATE 48000
 #define CHANNELS    2
@@ -48,16 +47,18 @@ int main(int argc, char** argv){
   SampleBuffer* samples = new SampleBuffer(BLOCKSIZE);
   if(argc > 1){
     const char* input_filename = argv[1];
-    WavHeader *wav_header = new WavHeader();
-    int16_t* data = NULL;
-    wavread(wav_header, input_filename, &data);
-    ASSERT(wav_header->num_channels == CHANNELS, "Incorrect number of channels in input file");
-    // ASSERT(wav_header->sample_rate == SAMPLE_RATE, "Incorrect sample rate in input file");
-    ASSERT(wav_header->bps == 16, "Incorrect number of bits per sample in input file");
-    const int channels = wav_header->num_channels;
-    const int len = wav_header->datachunk_size/(wav_header->fmtchunk_size/8);
-    int16_t* src = data;
-    int16_t* end = data+len;
+    uint8_t* data = NULL;
+    size_t len = fileread(input_filename, &data, 0);
+    data = (uint8_t*)malloc(len);
+    fileread(input_filename, &data, len);
+    WavFile wav = WavFile::create(data);
+    const int channels = wav.getNumberOfChannels();
+    // ASSERT(wav.getSampleRate() == SAMPLE_RATE, "Incorrect sample rate in input file");
+    ASSERT(channels == CHANNELS, "Incorrect number of channels in input file");
+    ASSERT(wav.getBitsPerSample() == 16, "Incorrect number of bits per sample in input file");
+    ASSERT(wav.isValid(), "Invalid wav");
+    int16_t* src = (int16_t*)wav.getData();
+    int16_t* end = src+wav.getNumberOfSamples();
     while(src+BLOCKSIZE <= end){
       samples->split16(src, BLOCKSIZE);
       processor.patch->processAudio(*samples);
@@ -66,10 +67,9 @@ int main(int argc, char** argv){
     }
     if(argc > 2){
       const char* output_filename = argv[2];
-      wavwrite(wav_header, output_filename, data);
+      filewrite(output_filename, data, len);
     }
     free(data);
-    delete wav_header;
   }else{
     processor.patch->processAudio(*samples);
   }
